+++ src/main/java/top/aiolife/record/api/IncomeController.java
@@ -85,7 +85,7 @@ public class IncomeController {
         
         // 获取收入类型字典数据
         List<SysDictDataEntity> dictDataList = sysDictService.getDictDataByDictType("income_type");
-        Map<Integer, String> dictMap = dictDataList.stream()
+        Map<Long, String> dictMap = dictDataList.stream()
                 .collect(Collectors.toMap(SysDictDataEntity::getDictCode, SysDictDataEntity::getDictLabel));
 
         // 按照年度汇总
@@ -112,7 +112,7 @@ public class IncomeController {
 
         // 获取收入类型字典数据
         List<SysDictDataEntity> dictDataList = sysDictService.getDictDataByDictType("income_type");
-        Map<Integer, String> dictMap = dictDataList.stream()
+        Map<Long, String> dictMap = dictDataList.stream()
                 .collect(Collectors.toMap(SysDictDataEntity::getDictCode, SysDictDataEntity::getDictLabel));
 
         // 按照年月汇总
